### 2023-01-16 #191 @whedon
[PRE REVIEW]: Course Materials for an Introduction to Data-Driven Chemistry
**Submitting author:** @ppxasjsm (<a href="http://orcid.org/0000-0001-7512-5252">Antonia S J S Mey</a>)
**Repository:** <a href="https://github.com/Edinburgh-Chemistry-Teaching/Data-driven-chemistry" target ="_blank">https://github.com/Edinburgh-Chemistry-Teaching/Data-driven-chemistry</a>
**Version:** v1.0.0.
**Managing EiC:** Jordan Gorzalski
