### 2023-01-17 #191 @labarba
Thank you for volunteering to edit this, @arm61 !!
I have looked over the paper and materials, and it looks like a good fit for JOSE. Check out this book for ideas of possible reviewers: [Teaching Programming across the Chemistry Curriculum](https://pubs.acs.org/isbn/9780841298194).
