+++ requirements.txt
@@ -2,7 +2,7 @@ pandas==1.4.2
 numpy==1.22.1
 matplotlib==3.5.2
 scipy==1.7.3
-ipywidgets==7.7.0
+ipywidgets==7.6.0
 jupyterlab==3.3.4
 nglview==3.0.3
 pint
