### 2023-01-17 #192 @whedon
[REVIEW]: Course Materials for an Introduction to Data-Driven Chemistry
**Submitting author:** @ppxasjsm (<a href="http://orcid.org/0000-0001-7512-5252">Antonia S J S Mey</a>)
**Repository:** <a href="https://github.com/Edinburgh-Chemistry-Teaching/Data-driven-chemistry" target ="_blank">https://github.com/Edinburgh-Chemistry-Teaching/Data-driven-chemistry</a>
**Version:** 1.1.0

Due to the challenges of the COVID-19 pandemic, JOSE is currently operating in a "reduced service mode".
