### 2023-01-17 #192 @arm61
Okay @lucydot and @ghutchis thanks for agreeing to review this interesting work! If you work through the checklist and if there is any problems/comments about the material, I would recommend opening an issue on [the material repository](https://github.com/Edinburgh-Chemistry-Teaching/Data-driven-chemistry) and the authors can sort them out. More information about the review guidelines can be found on the Open Journals documentation pages: https://openjournals.readthedocs.io/en/jose/reviewer_guidelines.html

If anyone has any questions just ping me on here!
