### 2023-01-18 Edinburgh-Chemistry-Teaching/Data-driven-chemistry#37 @lucydot
Suggested documentation improvements
Hello!

I'm working through the review criteria for your JOSS submission: https://github.com/openjournals/jose-reviews/issues/192

Very nicely organised project, and some clear documentation. Some smaller points below

- [x] This is a related course that you may want to reference (developed by researchers in Bath): https://pythoninchemistry.org/ch40208/introduction/about_this_book.html
- [x] I advise you include a citation for the Software Carpentry resources: https://software-carpentry.org/blog/2011/01/how-to-cite-software-carpentry.html
- [x] The reference section of the README is currently TBC: https://github.com/Edinburgh-Chemistry-Teaching/Data-driven-chemistry#reference
- [x] I think step-by-step installation instructions may be useful given audience is complete novice. Something like: https://github.com/lucydot/effmass#installation
- [x] I don't think its clear how others might adopt the module for their own teaching - this could be included in the README. I think it would be a case of simply cloning the repo - making it a template repo (https://docs.github.com/en/repositories/creating-and-managing-repositories/creating-a-template-repository) might be beneficial for those who may need multiple clones.
- [x] Contributing guidelines are fine, however if you *really* want to encourage people to contribute, you could maintain a list of areas you think need improvement, a wish list of things to improve (looks like you have something like this on Issues already; link to this and use of tags  - `good first issue` etc - might encourage those who need a little direction to begin with). You could also make clear that discussions (via issue thread or similar) about possible contributions are welcome.
