### 2023-01-18 #192 @lucydot
Hello!
Working through the checklist, will start today and will continue through till next week.

First issue (small points regarding documentation) raised here: https://github.com/Edinburgh-Chemistry-Teaching/Data-driven-chemistry/issues/37
