### 2023-01-18 Edinburgh-Chemistry-Teaching/Data-driven-chemistry#37 @ppxasjsm
Thanks for the above points! All excellent!  Just a couple of comments:

> This is a related course that you may want to reference (developed by researchers in Bath): https://pythoninchemistry.org/ch40208/introduction/about_this_book.html

Yes absolutely right! We even took inspiration from there! I'll add it.

> Reference section in README

This was just meant to be the final link to the JOSE publications, hence the TBC I am not sure if it is worth while duplicating information in multiple places.

>  I think step-by-step installation instructions may be useful given audience is complete novice.

I can add this, the idea was really to encourage novices to use Binder rather than a local install. There was some discussion around this and we weren't sure if we should commit to suggesting conda with a conda environment. Thoughts?

> Contributing guidelines are fine, however if you really want to encourage people to contribute

totally agree. I'll see if I can improve this!

>  I don't think its clear how others might adopt the module for their own teaching - this could be included in the README. I think it would be a case of simply cloning the repo - making it a template repo

sure!
