### 2023-01-18 Edinburgh-Chemistry-Teaching/Data-driven-chemistry#37 @lucydot
> This was just meant to be the final link to the JOSE publications, hence the TBC I am not sure if it is worth while duplicating information in multiple places.

Ah ok, yes that makes sense to leave empty until link put in

> I can add this, the idea was really to encourage novices to use Binder rather than a local install. There was some discussion around this and we weren't sure if we should commit to suggesting conda with a conda environment. Thoughts?

I've had trouble with Binder timing out on students, it is also restrictive in that students can't build on work they have done in previous sessions (as can't save the notebook for use later with the free binder service, as far as I know). Towards the end of the course hopefully students will be keen to code and so at some point will need to move to local install. Perhaps you could link here to step-by-step instructions elsewhere if you don't want to clog up the README with lower priority items.
