### 2023-01-18 Edinburgh-Chemistry-Teaching/Data-driven-chemistry#37 @ppxasjsm
> Perhaps you could link here to step-by-step instructions elsewhere if you don't want to clog up the README with lower priority items.

I will link to the carpentries resource for this and then give explicit info on how to install this particular environment. (NGLVIEW can be a bit tricky...I've had issues with it.)

I'll mark the References TBC as dealt with if that is ok.
