### 2023-01-18 4955b7a
references to training material and template repo
