### 2023-01-18 Edinburgh-Chemistry-Teaching/Data-driven-chemistry#37 @ppxasjsm
This commit 4955b7a87e7620548abf5d84b9cdb806e992193a addresses:

- Repo is a template repo.
- Information added to readme on how to build on this repo using the template (can be expanded if need be)
- added related repos of interest including book from bath

I'll add the bath book to the paper as well and make sure carpentries is cited appropiately in paper.md in a future commit.
