### 2023-01-24 Edinburgh-Chemistry-Teaching/Data-driven-chemistry#38 @lucydot
Other small suggestions
Hello,

Continuing with the JOSE review (https://github.com/openjournals/jose-reviews/issues/192) - now working through the "Pedagogy / Instructional design" section. I'll add more here as they come up.

Note: I haven't ran these through binder (see https://github.com/Edinburgh-Chemistry-Teaching/Data-driven-chemistry/issues/39) so reading the notebooks straight up.

- LO's for section 10 are written in different style (longer sentence) than the other parts (its a small point I know!)
- there are a few "FIXMEs" in various places (e.g. unit 2 and unit 3). I don't think this is necessarily a Bad Thing as resources are in a continuous form of improvement, but easier to use as-is if they are removed.
- on a similar point to the previous, I was a little confused as to whether this is designed as a clone-and-go project, or whether you are expecting instructors to edit before they can use - as there are some parts which seemed to be talking to instructors rather than students - e.g.  the Mentimetre section [here](https://github.com/Edinburgh-Chemistry-Teaching/Data-driven-chemistry/blob/main/Unit_01/Unit_01_problem_solving_II.ipynb). I expect most instructors would want to edit before teaching, but it might be nice to flag this up if it is required.

The Mentimetre functionality is great!
