### 2023-01-24 Edinburgh-Chemistry-Teaching/Data-driven-chemistry#39 @lucydot
Error launching binder
Hello,

Another one whilst working through the JOSE review (https://github.com/openjournals/jose-reviews/issues/192).

I'm unable to launch binder via the link given on the README (only tested for parts 1 and 2). The error is [here](https://gist.github.com/lucydot/e3b996b56bd8509a296979375f0fde8a).
