### 2023-01-24 #192 @lucydot
Hi @ppxasjsm a quick heads up that I've raised two other issues on the repo. I'll add to the "Other small suggestions" as I work through the course.
