### 2023-01-24 Edinburgh-Chemistry-Teaching/Data-driven-chemistry#39 @ppxasjsm
Ah yes, I may have broken this trying to fix NGLview, still trying to get this to work..fixing this is definitely on the to do list.
