### 2023-02-06 #192 @lucydot
@ppxasjsm I've updated https://github.com/Edinburgh-Chemistry-Teaching/Data-driven-chemistry/issues/38 with a few other bits.

I really like the problem solving introduction to the course and will use this for my own teaching. The Mentimetre integration is also great.

Note that I found it hard to judge the suitability of some of the questions for a UG chemistry curriculum as I studied physics, and am only familiar with a very particular part of chemistry (translation: I don't understand most of UG chemistry).

Some thoughts for the future (**not** acceptance blockers for this review):
- hosting this as an online book (using e.g. [quarto](https://quarto.org/)) might improve readability. There is the option to integrate binder buttons so that it is not at the sake of user interaction.
- would you consider including the slides that are referenced?
- as a lecturer, one of the things I most struggle with is designing/writing summative assessment. It'd be great to understand a little more about how summative assessment fits in with this course.
- I do see a bit of a disjoint between the parts of the course built from Software Carpentry resources (earlier chapters), and those that aren't (later chapters). The former have questions suitable for a more general audience, and explanations are split into smaller chunks (there are more headings). Exactly the same thing occured when I wrote a similar course for physicists, so I'm not really one to criticise. Ultimately, a course with domain specific examples throughout, and perfectly consistent writing style / lesson structure would be ideal. But I'd say thats a hypothetical end point to aim for rather than a requirement for a complete and effective course, which this appears to be. Clearly a lot of work has already gone into it as it is.
