### 2023-02-06 #192 @ppxasjsm
> -  hosting this as an online book (using e.g. [quarto](https://quarto.org/)) might improve readability. There is the option to integrate binder buttons so that it is not at the sake of user interaction.

Yes I've actually started looking into this, but then ran out of time to get an auto build for this for github actions...I can try and get this sorted for the review, but this will depend on when I have time to sort this out.

> would you consider including the slides that are referenced?

Yes absolutely! I'll add them.

> as a lecturer, one of the things I most struggle with is designing/writing summative assessment. It'd be great to understand a little more about how summative assessment fits in with this course.

Do you want this more generally or in the paper? We have two summative assessments after session 3 and 7 and then a mini workshop at the end that tests everything. I'd be happy to share one example of a summative assessment, but because we tend to reuse some of the material there are some restrictions around this.

> I do see a bit of a disjoint between the parts of the course built from Software Carpentry resources (earlier chapters), and those that aren't (later chapters). The former have questions suitable for a more general audience, and explanations are split into smaller chunks (there are more headings). Exactly the same thing occured when I wrote a similar course for physicists, so I'm not really one to criticise. Ultimately, a course with domain-specific examples throughout, and perfectly consistent writing style / lesson structure would be ideal. But I'd say thats a hypothetical end point to aim for rather than a requirement for a complete and effective course, which this appears to be. Clearly a lot of work has already gone into it as it is.

Yes I totally agree with your point. I think the smaller headings kind of make sense early on and may be harder to keep up later on but this can be streamlined. For next year, I've started putting together some extra practice material that is geared towards the topics for the first 3 lessons for at home practice, but this isn't complete yet. I've also found it difficult to find 'truely' domain specific examples for teaching syntax etc...
