### 2023-02-07 #192 @lucydot
Hi @ppxasjsm

> Yes I've actually started looking into this, but then ran out of time to get an auto build for this for github actions...I can try and get this sorted for the review, but this will depend on when I have time to sort this out.

I don't think it's required for the review, as the course can be followed clear enough as is.

> Do you want this more generally or in the paper? We have two summative assessments after session 3 and 7 and then a mini workshop at the end that tests everything. I'd be happy to share one example of a summative assessment, but because we tend to reuse some of the material there are some restrictions around this.

If you are able to share an example that would be great, but again not necessary for the review. Though it'd be great to have a chat with you about assessment sometime! That's outside the scope of this review though.
