### 2023-02-17 #192 @ppxasjsm
Hi @lucydot and @ghutchis, just a heads up that if you have any comments on the content of the course I have some capacity to make edits in the next couple of weeks to address any issues you might want to raise! Thanks :)
