### 2023-02-21 #192 @lucydot
Hi @ppxasjsm, I'll update the tick boxes following outcome on the discussions at https://github.com/Edinburgh-Chemistry-Teaching/Data-driven-chemistry/issues/39 , https://github.com/Edinburgh-Chemistry-Teaching/Data-driven-chemistry/issues/38 and https://github.com/Edinburgh-Chemistry-Teaching/Data-driven-chemistry/issues/37 .
