+++ requirements.txt
@@ -1,12 +1,12 @@
-pandas==1.4.2
-numpy==1.22.1
-matplotlib==3.5.2
-scipy==1.7.3
-ipywidgets==7.7.0
-jupyterlab==3.3.4
+pandas
+numpy
+matplotlib
+scipy
+ipywidgets==7.6.0
+jupyterlab
 nglview==3.0.3
 pint
 pytraj
-mendeleev==0.10.0
+mendeleev
 vpython
-py3Dmol==1.8.1
+py3Dmol
