### 2023-02-25 8fe6c80
adding colab and install instructions
