### 2023-02-25 Edinburgh-Chemistry-Teaching/Data-driven-chemistry#37 @ppxasjsm
Commit 8fe6c80 adds more detailed installation instructions
