### 2023-02-25 Edinburgh-Chemistry-Teaching/Data-driven-chemistry#38 @ppxasjsm
> LO's for section 10 are written in different style (longer sentence) than the other parts (its a small point I know!)

Commit f441cac should fix this.

> there are a few "FIXMEs" in various places (e.g. unit 2 and unit 3). I don't think this is necessarily a Bad Thing as resources are in a continuous form of improvement, but easier to use as-is if they are removed.

I'll double check, but all the FIXMEs should be tasks students would need to try and solve. The FXIME is just a dummy comment in an empty cell for them to add their code. If that is confusing I could swap this out to your solution here? Or just a blank cell? Thoughts?

>  [..] was a little confused as to whether this is designed as a clone-and-go project, [...]

Yeah good point. I'll make this a bit clearer. Will comment here when I've done that.
