+++ paper/paper.bib
@@ -233,3 +233,19 @@
   langid = {english},
   file = {/Users/toni_brain/Zotero/storage/YGWBRF5W/White et al. - 2022 - Data Carpentry for Biologists A semester long Dat.pdf;/Users/toni_brain/Zotero/storage/LX3SWXF6/jose.html}
 }
+@misc{software-carpentry,
+    author =       {Greg Wilson},
+    title =        {Software Carpentry web site},
+    howpublished = {http://software-carpentry.org},
+    accessed =     {Feburary 2023},
+    note =         {Main web site for Software Carpentry, replacing http://swc.scipy.org.}
+}
+@article{wilson-software-carpentry,
+    author =  {Greg Wilson},
+    title =   {Software Carpentry: Getting Scientists to Write Better Code by Making Them More Productive},
+    journal = {Computing in Science \& Engineering},
+    month =   {November--December},
+    year =    {2006},
+    note =    {Summarizes the what and why of Version 3 of the course.}
+}
+
