+++ overview.md
@@ -1,3 +1,24 @@
 # Welcome to Data-Driven Chemistry 
 
-This will contain a bit more information when finished. 
+## 1. Using the Material as a student for self study
+
+### If you are a student working on the material independently: 
+You can look at the images and try and answer the questions or safely ignore them entirely.
+
+
+## 2. Using the material as an instructor with a class of students
+
+### 2.1 Getting started
+The easiest way is by cloning the material and adapting it to your needs. This can be just using some partial material or expanding on the existing material. The best way to do this is by either [cloning](https://docs.github.com/en/repositories/creating-and-managing-repositories/cloning-a-repository) the repo and building up on it, or using the current repository as a [template repository](https://docs.github.com/en/repositories/creating-and-managing-repositories/creating-a-repository-from-a-template) for your own or your organisations GitHub account.
+
+### 2.2 Embedding Mentimeter
+
+
+### Mentimeters
+
+This is an example of the type of questions students get asked to answer via Mentimeter:
+
+Please see [mentimeter_example.ipynb](mentimeter_example.ipynb) for an example on how to embed Mentimeters into markdown cells. All following notebooks will have images of Mentimeter questions embedded.
+
+#### If you are an instructor:
+If you want to use in-class quizzes replace the images with your own questions using your [Mentimeter](https://www.mentimeter.com) account 
