### 2023-02-25 512c0ed
fixing f prints in helper function
