### 2023-02-27 a0828de
update paper to replace binder with colab
