### 2023-02-27 #192 @ppxasjsm
I decided to ditch binder. Updated paper to reflect only use of Colab.
