+++ README.md
@@ -8,18 +8,18 @@ Data-Driven Chemistry is a collection of learning material for an introductory P
 
 ## Course Content:
 
-| Unit | Content Summary                 | Binder    | Colab|
-|------|---------------------------------|-----------|------|
-| 1    |  An Introduction to algorithmic thinking and using Jupyter notebooks               |[![Binder](https://mybinder.org/badge_logo.svg)](https://mybinder.org/v2/gh/Edinburgh-Chemistry-Teaching/Data-driven-chemistry/HEAD?labpath=Unit_01%2FUnit_01_problem_solving_I.ipynb)| [![Unit_01](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/github/Edinburgh-Chemistry-Teaching/Data-driven-chemistry/blob/main/Unit_01/Unit_01_problem_solving_I.ipynb) |
-| 2    | Variables (int, float, string), Lists, Dictionaries and Tuples in Python|[![Binder](https://mybinder.org/badge_logo.svg)](https://mybinder.org/v2/gh/Edinburgh-Chemistry-Teaching/Data-driven-chemistry/HEAD?labpath=Unit_02%2FUnit_02_variables_I.ipynb) | [![Unit_02](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/github/Edinburgh-Chemistry-Teaching/Data-driven-chemistry/blob/main/Unit_02/Unit_02_variables_I.ipynb)|
-| 3    | Loops and conditional statements                                                 | [![Binder](https://mybinder.org/badge_logo.svg)](https://mybinder.org/v2/gh/Edinburgh-Chemistry-Teaching/Data-driven-chemistry/HEAD?labpath=Unit_03%2FUnit_03_loops_I.ipynb)| [![Unit_03](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/github/Edinburgh-Chemistry-Teaching/Data-driven-chemistry/blob/main/Unit_03/Unit_03_loops_I.ipynb)|
-| 4    | Functions and basic Input/Output                                       | [![Binder](https://mybinder.org/badge_logo.svg)](https://mybinder.org/v2/gh/Edinburgh-Chemistry-Teaching/Data-driven-chemistry/HEAD?labpath=Unit_04%2FUnit_04_functions_I.ipynb)| [![Unit_04](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/github/Edinburgh-Chemistry-Teaching/Data-driven-chemistry/blob/main/Unit_04/Unit_04_functions_I.ipynb)|
-| 5    | An introduction to plotting, using units and statistical analysis                     | [![Binder](https://mybinder.org/badge_logo.svg)](https://mybinder.org/v2/gh/Edinburgh-Chemistry-Teaching/Data-driven-chemistry/HEAD?labpath=Unit_05%2FUnit_05_I_numerical_data.ipynb) | [![Unit_05](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/github/Edinburgh-Chemistry-Teaching/Data-driven-chemistry/blob/main/Unit_05/Unit_05_I_numerical_data.ipynb)|
[…]
