### 2023-02-27 Edinburgh-Chemistry-Teaching/Data-driven-chemistry#39 @ppxasjsm
I couldn't get Binder to work in a fast way (I saw you can do prebuilds of docker containers with github actions, but it seemed like faff to figure out how to do that). Also NglView was a pain, so I migrated to Colab. This should be all up and running now.
