### 2023-02-27 #192 @ppxasjsm
> https://github.com/Edinburgh-Chemistry-Teaching/Data-driven-chemistry/issues/39

Fixed this. Now completely on colab. There is a bit of extra code for pulling data in, but should be useable.
