+++ overview.md
@@ -1,33 +1,44 @@
 # Welcome to Data-Driven Chemistry 
 
-## 1. Using the Material as a student for self study
+Data-Driven Chemistry is a course aimed at 2nd year UG students in Chemistry in Scotland, but we have made the material available for everyone. It can be used for self study, or as a template for other instructors to duplicate or build on. Below we summarise how to best interact with the material for these two cases.  
 
-### If you are a student working on the material independently: 
-You can look at the images and try and answer the questions or safely ignore them entirely.
+## 1. Using the material as a student for self study
+
+If you are a student working on the material independently, you can just work your way through each notebook.
+
+### 1.1 Working in Colab
+The easiest way to do this is by using the provided Colab links. They should work out of the box with all necessarily packages installed. If you run into trouble sometimes restarting the runtime environment may simply solve this issue. 
+
+### 1.2 Using a local installation
+If you want to use a local isntallation, please follow the instructions on how to install necessary dependencies provided in the main [README.md](README.md) file. 
+
+In which ever way you load the notebooks, you will be able to execute cells and follow the explanations and instructions. Tasks are marked as such and will each provide a template solution in a drop down box. 
+
+### 1.3 Metimeter
+Some cells currently show images of Mentimeter quizzes. These are used for in-class quizzes. You can look at them and think of an answer you may give, but are not essential the course. They are meant more as a guide for instructors to add should they so desire. 
+
+### 1.4 Extra material
+Some units contain extra notebooks for additional practice. These are less well curated than the core units and can be used for the curious, but do not contain essential material. 
 
 
 ## 2. Using the material as an instructor with a class of students
 
+The material can be used in any way you like. Below we outline how to use it as is and incorporate additional features we have used in the course at Edinburgh. However, any part of it can be reused and adapted to your needs. If you tell us about how you found material useful we would appreciate this alot!
+
 ### 2.1 Getting started
-The easiest way is by cloning the material and adapting it to your needs. This can be just using some partial material or expanding on the existing material. The best way to do this is by either [cloning](https://docs.github.com/en/repositories/creating-and-managing-repositories/cloning-a-repository) the repo and building up on it, or using the current repository as a [template repository](https://docs.github.com/en/repositories/creating-and-managing-repositories/creating-a-repository-from-a-template) for your own or your organisations GitHub account.
+The best way to get started is to either [cloning](https://docs.github.com/en/repositories/creating-and-managing-repositories/cloning-a-repository) the repo and building up on it, or using the current repository as a [template repository](https://docs.github.com/en/repositories/creating-and-managing-repositories/creating-a-repository-from-a-template) for your own or your organisations GitHub account.
 
-### 2.2 Embedding Mentimeter
-
-
-### Mentimeters
-
-This is an example of the type of questions students get asked to answer via Mentimeter:
-
-Please see [mentimeter_example.ipynb](mentimeter_example.ipynb) for an example on how to embed Mentimeters into markdown cells. All following notebooks will have images of Mentimeter questions embedded.
-
-#### If you are an instructor:
-If you want to use in-class quizzes replace the images with your own questions using your [Mentimeter](https://www.mentimeter.com) account
+### 2.2 Preparing files for Colab or running locally
[…]
