### 2023-03-03 Edinburgh-Chemistry-Teaching/Data-driven-chemistry#37 @ppxasjsm
I've expanded on how to contribute with this: d01023a. I don't think I want to add more at this stage and see how it goes.
