### 2023-03-03 Edinburgh-Chemistry-Teaching/Data-driven-chemistry#37 @ppxasjsm
I think all points are now addressed here, so closing the issue.
