### 2023-03-09 #192 @lucydot
@ppxasjsm I've just read through the three issues, and I agree all have been addressed. I'm happy to recommend this work for publication - and I plan to adapt some of this for use in my own teaching :star:
