### 2023-03-09 #192 @ppxasjsm
Thank you @lucydot. Glad you think the material may also be useful to you!
