### 2023-03-21 Edinburgh-Chemistry-Teaching/Data-driven-chemistry#50 @ghutchis
Discussing Bootstrapping / error in the Median, skew
I'm not sure whether this is better suited for unit 5 or 6. Since non-Gaussian distributions come up fairly commonly in chemistry (e.g., Boltzmann) I think it's important to emphasize:
- you may want to check the skew using `scipy.stats.skew`
- the median may be a better "central metric" than the mean (e.g., income distribution if Jeff Bezos comes into the room)
- to estimate the std. dev. of the median, you can't use `np.std()`
- bootstrapping is a nice, simple way to get confidence intervals on median or other properties, e.g.

https://carpentries-incubator.github.io/machine-learning-novice-python/07-bootstrapping/index.html
https://github.com/ghutchis/chem1000/blob/main/lectures/09b-prob-stats.ipynb
