### 2023-03-21 #192 @ghutchis
Overall, I think you've fixed a lot of things I noted on my first assessment. I think my remaining issue would be:
https://github.com/Edinburgh-Chemistry-Teaching/Data-driven-chemistry/issues/50
