### 2023-03-23 4f8bad5
some additional info on skew lesson 7
