### 2023-03-23 Edinburgh-Chemistry-Teaching/Data-driven-chemistry#50 @ppxasjsm
I ended up putting it into Unit_07. I remember that when teaching 6 I just had enough material without overrunning, but unit 7 had a bit more breathing time. It fits well well after the plotting distributions recap as well.
Commit 4f8bad5 implements our proposed final version for this. I have not included boostrapping as this is something that is taught in a different course in a later year. The general idea is to avoid teaching concepts in chemistry and programming and tend to rely on the students' knowledge of stats, etc., taught in their first year in a maths course.
