+++ README.md
@@ -66,16 +66,16 @@ Authors are, in alphabetical order:
 - Hannah Pollak (HP)
 - Rafal Szabla (RS)
 
-JC, VE, CLH, and ASJSM have been teaching this course at the University of Edinburgh since the academic year 2021/22. JJG and AP have been course demonstrators. RS taught one unit and created content for it in 2020/21, when the course was run in a shortened form as a replacement for physical chemistry laboratory practicals during the pandemic. MTD shared and adapted materials developed in 2018 for his course at Durham University aimed at chemistry research students, and made some additional contributions.
+JC, VE, CLH, and ASJSM have been teaching this course at the University of Edinburgh since the academic year 2021/22. JJG and HP have been course demonstrators. RS taught one unit and created content for it in 2020/21, when the course was run in a shortened form as a replacement for physical chemistry laboratory practicals during the pandemic. MTD shared and adapted materials developed in 2018 for his course at Durham University aimed at chemistry research students, and made some additional contributions.
 
 Specific contributions by each author are as follows.
-- JC: Created the material for `Unit_01`, `Unit_10` and the helper_functions, gave feedback on other materials, and helped edit the manuscript.    
-- MTD: Contributed material to `Unit_03`, `Unit_05`, `Unit_07` and `Unit_08`, and helped edit the manuscript.    
+- JC: Created the material for `Unit_01`, `Unit_10`, and the helper_functions, gave feedback on other materials, and helped edit the manuscript.    
+- MTD: Contributed material to `Unit_03`, `Unit_05`, `Unit_07`, and `Unit_08`, and helped edit the manuscript.    
 - VE: Created material for `Unit_05` and `Unit_08`, contributed to `Unit_06`, and helped edit the manuscript.   
 - JJG: Contributed material to `Unit_09`, and helped edit the manuscript.   
 - CLH: Created material for `Unit_03` and `Unit_04`, and helped edit the manuscript.    
 - ASJSM: Created material for `Unit_02`, `Unit_06`, `Unit_07`, and `Unit_09`, provided feedback and small contributions to most other units, and wrote the manuscript.   
-- HP: Contributed material to `Unit_05` and  `Unit_08`, and helped edit the manuscript.   
+- HP: Contributed material to `Unit_05` and `Unit_08`, and helped edit the manuscript.   
 - RS: Created the material for molecular geometries forming part of `Unit_06` and gave feedback on the manuscript.   
 
 ## Acknowledgements
