### 2023-03-27 #192 @arm61
@ppxasjsm two things in the references for resolve:
- The author for `pandas` renders pretty weird at the moment (I think cause of the curlys) can you have a look.
![Screenshot 2023-03-27 at 11 07 48](https://user-images.githubusercontent.com/15774281/227895910-ef7186ed-c167-4553-8355-7b71b67bd6cb.png)
- The DOI is missing for the software carpentry stuff. The DOI that the bot (who we shall not name) has found seems like the right one.
