+++ paper/paper.bib
@@ -186,9 +186,9 @@
 }
 
 @misc{team2022pandasdev,
-  title = {Pandas-Dev/Pandas: {{Pandas}}},
+  title = {Pandas-Dev/Pandas: Pandas},
   shorttitle = {Pandas-Dev/Pandas},
-  author = {pandas development {team}, The},
+  author = {The pandas development team},
   year = {2022},
   month = nov,
   doi = {10.5281/zenodo.7344967},
@@ -246,6 +246,7 @@
     journal = {Computing in Science \& Engineering},
     month =   {November--December},
     year =    {2006},
-    note =    {Summarizes the what and why of Version 3 of the course.}
+    note =    {Summarizes the what and why of Version 3 of the course.},
+    doi =     {10.1109/mcse.2006.122}
 }
 
