### 2023-03-27 #192 @ppxasjsm
Thank you @arm61! I am working on the DOI/references stuff.
