+++ paper/paper.bib
@@ -185,16 +185,16 @@
   langid = {english}
 }
 
-@misc{team2022pandasdev,
-  title = {Pandas-Dev/Pandas: Pandas},
-  shorttitle = {Pandas-Dev/Pandas},
-  author = {The pandas development team},
-  year = {2022},
-  month = nov,
-  doi = {10.5281/zenodo.7344967},
-  abstract = {This is a patch release in the 1.5.x series and includes some regression and bug fixes. We recommend that all users upgrade to this version. See the full whatsnew for a list of all the changes. The release will be available on the defaults and conda-forge channels: conda install pandas Or via PyPI: python3 -m pip install --upgrade pandas Please report any issues with the release on the pandas issue tracker. Thanks to all the contributors who made this release possible.},
-  howpublished = {Zenodo},
-  file = {/Users/toni_brain/Zotero/storage/9RCWPU6J/7344967.html}
+@software{team2022pandasdev,
+  author       = {The pandas development team},
+  title        = {pandas-dev/pandas: Pandas},
+  month        = nov,
+  year         = 2022,
+  note         = {If you use this software, please cite it as below.},
+  publisher    = {Zenodo},
+  version      = {v1.5.2},
+  doi          = {10.5281/zenodo.7344967},
+  url          = {https://doi.org/10.5281/zenodo.7344967}
 }
 
 @article{virtanen2020scipy,
