### 2023-03-27 #192 @arm61
My only comment from proof reading (note some mild dyslexia means that proof reading probably isn't my best skill) is that on line 48 of the paper (https://github.com/Edinburgh-Chemistry-Teaching/Data-driven-chemistry/blob/9e37bde609c11a2df932c63cc3be07e08deb2a44/paper/paper.md?plain=1#L48) you should use the possessive form of "Software Carpentery" (i.e. "Software Carpentry's").
