+++ paper/paper.bib
@@ -186,7 +186,7 @@
 }
 
 @software{team2022pandasdev,
-  author       = {The pandas development team},
+  author       = {The pandas development team,{}},
   title        = {pandas-dev/pandas: Pandas},
   month        = nov,
   year         = 2022,
