### 2023-03-27 #192 @ppxasjsm
I think I have now fixed the pandas reference. Can you let me know if you are happy with this? I'll send out the current version for final proof reading comments from my co-authors.
