### 2023-03-27 #192 @arm61
For the pandas reference I think something like
```
  author       = {{The pandas development team}},
```
might remove the weird capital T (cause LaTeX thinks that "The" is someones first name.
