### 2023-03-27 #192 @ppxasjsm
I added a blank first name like this:

```
author       = {The pandas development team, {}},
```
