### 2023-03-27 #192 @arm61
Might be getting into a bit of a rabbit hole, but that now adds a full stop at the end of the author which shouldn't be there (the full stop in the other references is to indicate the abbreviation of the author's first names). I will have a look into the reference style that JOSE is using.
