### 2023-03-27 #192 @arm61
According to this issue (https://github.com/openjournals/joss/issues/230, for JOSS but JOSE uses the same bot) the format is author-year (https://www.overleaf.com/learn/latex/Biblatex_bibliography_styles) so there shouldn't be a full stop between the author name and the bracketed year (unless the author first name is abbreviated). Can you try wrapping everything in curlys and see if that fixes it?
