### 2023-03-27 bd93701
another try with pandas
