### 2023-03-27 #192 @ppxasjsm
Aren't there full stops after all the author names? To me it doesn't look different to other references, but happy to hear about an improved solution.
