### 2023-03-27 #192 @arm61
There are, but those full stops indicate the abbreviation of the name i.e. Andrew -> A., however I have since found out that for “group authors” there is a full stop after (https://apastyle.apa.org/style-grammar-guidelines/references/examples/webpage-website-references). So I guess it is good as is! Sorry for the faff (I am still learning!).
