### 2023-03-27 #192 @arm61
The next step is minting a DOI with Zenodo but we should wait until your coauthors have proof read before we go to that step.
