### 2023-03-27 6643189
Update paper.md
small edit to authors contributions
