+++ paper/paper.md
@@ -44,29 +44,30 @@ bibliography: paper.bib
 
 # Summary
 
-Data-Driven Chemistry is a course aimed at undergraduate students in chemistry with no prior knowledge of programming and programmatic data analysis. It is designed as a 10-week-long course,^[This is a 20 SCQF credit course including assessments, roughly equivalent to 10 ECTS or 5 US credits] introducing Python programming and its usage in data analysis typically required for a chemistry degree. The course consists of 10 Units, designed to be used in a blended learning environment of live coding and explanations, followed by a set of in-course tasks to be solved individually or through pair programming.
-Generally, we aimed to follow Software Carpentry [@wilson-software-carpentry] teaching philosophy as closely as possible for each Unit. While, in principle, this course is designed to be taught as a whole, each Unit will cover a set of self-contained topics. The idea behind the modular design is to make the reuse and mix-and-match with other courses as easy as possible.
+Data-Driven Chemistry is a course aimed at undergraduate students in chemistry with no prior knowledge of programming and programmatic data analysis. It is designed as a 10-week-long course,^[This is a 20 SCQF credit course including assessments, roughly equivalent to 10 ECTS or 5 US credits] introducing Python programming and its usage in data analysis typically required for a chemistry degree. The course consists of 10 units designed to be used in a blended learning environment of live coding and explanations, followed by a set of in-course tasks to be solved individually or through pair programming.
+
+In general, we aimed to follow the teaching philosophy of Software Carpentry [@wilson-software-carpentry] as closely as possible for each unit. While this course is designed to be taught as a whole, each unit will cover a set of self-contained topics. The modular design is intended to make it easy to reuse and mix-and-match with other courses.
+
 The original material is hosted online using the [Noteable](https://noteable.edina.ac.uk/) service provided by the University of Edinburgh, but to make it available outside of the University of Edinburgh, we have migrated the content to [Google CoLab](https://colab.research.google.com). The online hosting on CoLab will enable self-guided learning, as well as classroom-based learning, ensuring that the learning is not only limited to a university classroom setting.
 
 # Statement of Need
 
-The modern world is digital, allowing for upscaling and automation of chemical processes through robots, but also fast production of large-scale datasets. Data analyses carried out with Graphical User Interfaces (GUIs) or spreadsheet-based tools are often limited in robustness, speed, and reproducibility. Programmatic solutions fare much better in this context, but programming is not typically taught as a skill across chemistry degrees, unlike in physics, mathematics, or even biology [@white2022data]. Both the Royal Society of Chemistry [@employability] and the American Chemical Society [@neiles2020professional] have identified good computational skills to be a key for graduate employability [@hill2019undergraduate]. Our course is designed to address this gap in the undergraduate chemistry curriculum at the University of Edinburgh and, thus, to ensure that chemistry graduates remain competitive with other STEM graduates. The material is made available as open source, with the hope that it may be used in other educational settings.
+The modern world is digital, allowing for upscaling and automation of chemical processes through robots, but also enabling the fast production of large-scale datasets. Data analyses carried out with Graphical User Interfaces (GUIs) or spreadsheet-based tools are often limited in robustness, speed, and reproducibility. Programmatic solutions fare much better in this context, but programming is not typically taught as a skill across chemistry degrees, unlike in physics, mathematics, or even biology [@white2022data]. Both the Royal Society of Chemistry [@employability] and the American Chemical Society [@neiles2020professional] have identified good computational skills as key for graduate employability [@hill2019undergraduate]. Our course is designed to address this gap in the undergraduate chemistry curriculum at the University of Edinburgh and to ensure that chemistry graduates remain competitive with other STEM graduates. The material is made available as open source, with the hope that it may be used in other educational settings.
 
[…]
