### 2023-03-29 #192 @ppxasjsm
I have pumped the version to 1.1.0 for the changes that have been made since the submission.
The DOI from Zenodo is: 10.5281/zenodo.7782433
