### 2023-05-20 #192 @labarba
Please merge my PR with minor edits to the paper: https://github.com/Edinburgh-Chemistry-Teaching/Data-driven-chemistry/pull/52
…and double check the figure caption, which isn't showing for me in the [preview service](https://whedon.theoj.org), but should! (I'm not sure what is going on there.)

I checked everything else, and will publish as soon as you merge and fix the caption as needed.
