### 2023-05-20 #192 @ppxasjsm
Hi @labarba I have merged edits and have figured out the figure caption (it was a missing new line) and things are now displaying ok for me.
