### 2023-05-20 #192 @labarba
One last thing! Can you check if adding curly brackets will fix the way "Project Jupyter" is appearing as an author in the nbgrader paper? It shows as "Jupyter, P." and that is not quite what we want...
