+++ paper/paper.bib
@@ -90,7 +90,7 @@
 @article{jupyter2019nbgrader,
   title = {Nbgrader: {{A Tool}} for {{Creating}} and {{Grading Assignments}} in the {{Jupyter Notebook}}},
   shorttitle = {Nbgrader},
-  author = {Jupyter, Project and Blank, Douglas and Bourgin, David and Brown, Alexander and Bussonnier, Matthias and Frederic, Jonathan and Granger, Brian and Griffiths, Thomas L. and Hamrick, Jessica and Kelley, Kyle and Pacer, M. and Page, Logan and P{\'e}rez, Fernando and {Ragan-Kelley}, Benjamin and Suchow, Jordan W. and Willing, Carol},
+  author = {{Jupyter Project} and Blank, Douglas and Bourgin, David and Brown, Alexander and Bussonnier, Matthias and Frederic, Jonathan and Granger, Brian and Griffiths, Thomas L. and Hamrick, Jessica and Kelley, Kyle and Pacer, M. and Page, Logan and P{\'e}rez, Fernando and {Ragan-Kelley}, Benjamin and Suchow, Jordan W. and Willing, Carol},
   year = {2019},
   month = jan,
   journal = {Journal of Open Source Education},
