### 2023-05-20 #192 @ppxasjsm
It now says Jupyter Project et al. is this acceptable?
