### 2023-05-20 #192 @labarba
Congratulations, @ppxasjsm, your JOSE paper is now published! 🚀

Huge thanks to our Editor: @arm61 and Reviewers: @lucydot, @ghutchis — your contributions make all of this possible 🙏
