### 2023-05-20 #192 @labarba
@openjournals/dev — Help requested! We have published a paper, and post-publication the author has indicated that there's an error in the ORCID. Are you able to make a fix to an already deposited article?

@ppxasjsm Please note here what the correct ORCID should be.
